[Wallet] Fix handling deep links (#5591)
### Description

Handling deep links sometimes fails, depending on whether the app if foreground/background, and the process is running or not.

For ios the issue was that the deep link was not immediately available during app init, so using the url listener in the App component fixes it.

For android it was kind of the opposite problem. It only worked if the app was not running, because if you exit the app with the back button we would remove the url listener. When the activity was started again from the intent appInit was not getting called since the app was already running. So, checking for the url in the componentDidMount method fixes it.

### Other changes

Also wrote an e2e test to expose the issues and to make sure they don't break in the future.

### Tested

Used adb to send an intent with the deep link to the app in various states.

### Related issues

- Fixes #4946 

### Backwards compatibility

This change is backwards compatible.