Use hard-coded gas values in the verifications flow (#5679)
### Description

Especially on high latency and slow network connections, gas estimation can regularly take a very long time, often 10-20 seconds, and represents a point of failure if network conditions are degraded.

In this PR, gas values are set statically based on the gas usage of representative transactions. This eliminates the need for gas estimation and will increase performance and reliability of the verifications flow. I've chosen to use 1.5x the highest observed gas usage in representative transactions.

One risk of this change is that it makes the flow fragile to significant future changes to the Attestations flow. If for some reason in the future, the Attestations smart contract requires a significant amount more gas to run the same functions, old client will not be compatible. I am proposing this change now because I see that changes that increase the gas cost by more than 1.5x, without breaking interface, are fairly unlikely and will certainly not be done regularly.

### Tested

Tested by running through the verifications flow on Alfajores. Including on my WiFi and on simulated 3G connectivity.

Note:
When testing with simulated 3G connectivity, the flow was completed up to having selected issuers in a little over 1 minute (compared to about 45 seconds on WiFi) but failed on first try and took about 2 minutes 15 seconds on the second try when using `master`. 

### Backwards compatibility

100%