[Wallet] Retry if geth fails to init because of no internet (#5700)
### Description

We explicitly crash the app if geth fails to initialize on startup. There are a few different cases how that can happen but one way to cause it is to start the app in airplane mode. This causes a network failure which immediately sends the user to the "Oops" screen.

The fix is to implement a retry. Of course this can go on forever if the network connection never comes back, so we could implement a limited number of retries. I don't think that is a good design though as we should never see the Oops screen or intentionally kill the app. Instead we should provide hints to the user as to how they might fix it. In this case we do have a Connecting banner, but we could do more such as showing a popup maybe. In general it is hard to tell why the app cannot connect. For example there might be a port block and the network is fine otherwise. I have some ideas here, such as using google firebase to check network connectivity, but that's for another time.

### Other changes

There are other geth failure modes such as when the timeout hits. This is a bit trickier to fix because we would need to cancel the initial init and start a new one. I will leave that for another PR after discussing with @jeanregisser how to implement that.

### Tested

Steps to reproduce on simulator or real device (TODO: write an e2e test):
1. kill app
2. turn off wifi if simulator, or go to airplane mode on device
3. open app

### Related issues

- Fixes #4867

### Backwards compatibility

Yes.