[Wallet] Remove force unlock from initial onboarding flow (#5734)
### Description

https://github.com/celo-org/celo-monorepo/pull/5656 has introduced mandatory enter of PIN code while onboarding flow to prevent account unlock timeout issue.

This PR would disable it specifically to the initial onboarding (we hope that 10 minutes since user sets the PIN would be enough to finish onboarding initially) and keep it if user comes back to verification later. 

### Caveat
The only one I see is when user goes to verification from Home screen, enters PIN, press Start/Resume, press back, user again needs to enter a PIN code. Not sure how often this is happening and to which extent it would ruin the UX.

### Tested

Initial onboarding with verification and return to verification flow later.

### Backwards compatibility

Yes