MTWD remove state mapping and permissions (#5683)
### Description

Based on multiple discussions we came to the conclusion that we should remove the on-chain mapping of EOA->MTW from the MetaTransactionWalletDeployer. This mapping can grow stale when recovery enables a MTW to switch signer.

By removing that state it also makes sense to make the wallet deployment permissionless as there's no state.