CeloTool simplify contexts (#5883)
### Description

Simplify the context management. Instead of having to worry about the service that you're deploying to, we're just including the service name in the context label. This makes it straightforward if we want to deploy full nodes to the Oracle context for example.

### Other changes

Increase Komenci timeout from five seconds to seven seconds.

### Tested

_An explanation of how the changes were tested or an explanation as to why they don't need to be._

### Related issues

- Fixes #[issue number here]

### Backwards compatibility

_Brief explanation of why these changes are/are not backwards compatible._