[Wallet] Upgrade react-navigation libraries (#5896)
### Description

This seems to fix the top crash we are currently experiencing.
Thanks @cmcewen for smelling the right fix here. We should also start a PR for upgrading all the other react packages, but I wanted to do this one as a smaller PR to make sure it fixes it and doesn't introduce anything new.

TODO: fix the any types that are added by asking @jeanregisser 

### Other changes

Added a new e2e test for opening the app using a dappkit url, which is the common source of crashes.

### Tested

Manually by opening the app with a deep link. Committed the script used also.

### Related issues

- Fixes #5524 

### Backwards compatibility

Yes.