Use migrationsConfig from release branch (#5687)
### Description

Adjusts `make-release` to use the migrationsConfig from the git reference we are deploying

### Related issues

- Fixes #5590 

### Backwards compatibility

Question: Just wanted to check that using the migrations config from celo-core-contracts-v1.rc1 is ok to do, even though it specifies FixidityLib as a dependency. As far as I can tell, it’s just gonna result in useless FixidityLib links or will it prevent an inline?