Fix revoked accounts appearing as verified (#5988)
### Description
When an account is revoked using the `revoke` function, the account is no longer associated with the given identifier but the attestation stats for that account remain unchanged. In our current verification flow, we only check attestation stats when determining if a given account should be considered as verified. This was causing us to erroneously consider revoked accounts to still be verified.

This PR adds an explicit check to the verification flow for whether an account is associated with an identifier when determining if that account is verified. If the account is not associated with the expected identifier, it is not considered verified regardless of how many successful attestations it has.

Additionally, we now have a warning modal on the `revoke` button that tells users that revoked accounts will not be able to be reverified.

### Other changes
- Fixed a bug that was preventing escrow withdrawals to fail during user funded verification
- Added `bundle exec pod install` to the `build:wallet` script

### Tested
Yes

### Related issues
- Fixes #5986 
- Fixes #5987

### Backwards compatibility
Yes