Show error modal when feeless verification flow encounters an error (#6275)
### Description
Fixed a small bug with regards to how errors are displayed during feeless vs. non-feeless attestations that was introduced in [this PR](https://github.com/celo-org/celo-monorepo/pull/5988). A flaw in logic for when `feelessIsActive` was toggled was causing it to prematurely turn to `false` when the feeless flow encountered an error, resulting in the error modal not appearing. The solution is to turn `feelessIsActive` to `false` when the verification flow is canceled (including cancelling due to an error), rather than simply when an error is encountered.

### Other changes
N/A

### Tested
Yes

### Related issues
- Fixes #6215

### Backwards compatibility
Yes