[Wallet] Try to fix flaky e2e test that reopens the app with a deep link (#6305)
### Description

A really flaky test is the deep link one, which closes and reopens the app with a deep link. From the screenshots it looks like the app does open properly but it does not navigate to the desired handler (handle send or dappkit).

I think there could be a timing issue which is why I am trying a couple sleeps to see if that improves flakiness.

Locally the e2e test runs fine for me but maybe on CI the emulator is slower which can cause timing failures.

### Other changes

None

### Tested

e2e

### Related issues

- Fixes #[issue number here]

### Backwards compatibility

Yes.