Fix param order on account:new internal call (#6319)
### Description
Why:
* The function definition on https://github.com/celo-org/celo-monorepo/blob/672378a328ec2c1e67271026f51bbe0236407ca8/packages/sdk/utils/src/account.ts#L75 is expecting `changeIndex` and then `addressIndex`, yet on https://github.com/celo-org/celo-monorepo/blob/672378a328ec2c1e67271026f51bbe0236407ca8/packages/cli/src/commands/account/new.ts#L114 it is called with `addressIndex` first and then `changeIndex`

This change addresses this issue by:
* Passing changeIndex first and then addressIndex

TBH, I'm not sure how breaking this is, but it looks very breaking. Just decided to open this because it looks like a simple change.
I don't have time now but this should also be verified by tests. I can add them later if you'd like.