Update docs CLI docs to not give running a full node as a prerequisite (#6374)
### Description

Users on Discord have been having difficulties with the instructions on recovering Celo from an Ethereum address.  The primary issue is that the instructions point to the CLI introduction page in the docs, which lists running a full node as a prerequisite to using the CLI.  This isn't right, and running a full node is technically quite challenging (and would take a while for the node to sync even if you did do it).  So I updated the CLI docs to make clear that running a full node is optional and that using Forno is the easier way to go.

### Other changes

In passing, I noticed the Ethereum recovery page was referencing the old url for mainnet Forno (https://rc1-forno.celo-testnet.org).  I updated it there and elsewhere in the docs to instead use the new url (https://forno.celo.org/).  There are a couple of other references to the old url:  https://github.com/celo-org/celo-monorepo/blob/73e1fd0831f1975e8ff0683d64b46ec946b922db/packages/komencikit/examples/attestation-flow.ts#L83 and https://github.com/celo-org/celo-monorepo/blob/73e1fd0831f1975e8ff0683d64b46ec946b922db/packages/notification-service/config/config.mainnet.env#L10 , but these are not in the docs so I wasn't sure whether I should update them in this PR or not.

### Tested

Verified that the forno url and the cross-links to other doc pages are correct.