[Wallet] Add dynamic notification to tray from Firebase (#6391)
### Description

Allows adding a notification to the `NotificationBox` dynamically from Firebase. The CTA opens a `WebView` with an URI received from Firebase.

<img width="357" alt="Screen Shot 2021-01-06 at 02 58 33" src="https://user-images.githubusercontent.com/6062888/103734601-12f60300-4fcb-11eb-90cb-b1c14d5124f5.png">

### Other changes

Added a `darkMode` option to the notifications on the `NotificationBox`. Dynamic notifications only look good on dark mode for now since we only have one icon, but adding a different icon for normal mode could fix that in the future.

### Tested

Yes

### Related issues

- Fixes #6137

### Backwards compatibility

N/A