[Wallet] Add deep links for Cash In screen (#6399)
### Description

Added the ability to send deep links to the cash in screen. The first use case of this will be a web page shown on a WebView which will have a button that will redirect to the cash in screen. The WebView has to redirect to `celo://wallet/cashIn` which will redirect to the `FiatExchangeOptions` screen.

### Tested

I used a custom HTML:
```
<html>
<body>
<a href="celo://wallet/cashIn">Click me</a>
</body>
</html>
```
and clicked on the `Click me` link. The app redirected me to the `FiatExchangeOptions` screen.

### Related issues

- Fixes #6398