[Wallet] Allow opening WebView or external browser from push notification (#6405)
### Description

We allow opening a web page from a push notification right now using an `ou` field on the notification. This field will now open a WebView instead. If we want to open an external webpage we will need to add a new field called `openExternal` with a `true` value in addition to the `ou` field.

### Tested

Sent push notifications with `ou` and `openExternal ` and only `ou` and checked that it behaved as expected.

### Related issues

- Fixes #6404

### Backwards compatibility

See the description