Show the international format when displaying phone numbers in Valora (#6350)
### Description
This PR changes it so that we show the international phone number when onboarding. Previously, we were attempting to do a combination of domestic and international when displaying, which resulted in the confusing outcomes detailed in the issues below.

### Other changes
Removed the international prefix from the drop down selector with the flag emojis so it was clear the prefix was not being double entered. Updated the GoogleLib package, which addresses some of the country specific issues we were seeing.

### Tested
Yes

### Related issues
- Fixes #5260
- Fixes #5261

### Backwards compatibility
Yes


![IMG_0909](https://user-images.githubusercontent.com/40568900/103042848-c40e8f00-452f-11eb-812a-b7a5894426c9.PNG)

![IMG_0910](https://user-images.githubusercontent.com/40568900/103042847-c40e8f00-452f-11eb-9a5f-b5d8b9f7f9a2.PNG)

![IMG_0911](https://user-images.githubusercontent.com/40568900/103042846-c2dd6200-452f-11eb-9e2e-62285ebdbe59.PNG)