Update Governance voting docs and add reminder for Ledger users (#6452)
### Description

Suggested by @human on Discord, this PR adds a brief reminder to include the nessasary flags when
using a Ledger and `celocli` to vote on Governance.

### Other changes

This PR also includes some minor updates to the documentation including changing the suggestion from
using the `governance:view` command to `governance:show` instead, and making the placholders used in
the example commands consistent.

### Tested

Manually tested the suggested commands