[Wallet] Fix navigation init (#6465)
### Description

Trying to fix the e2e test that fails on the deep link. Looking at the logs I see a navigation failure like this:

> [Wed Jan 13 2021 08:29:06.736]  INFO     NavigationService@navigate :: Navigation failure: Error: The 'navigation' object hasn't been initialized yet. This might happen if you don't have a navigator mounted, or if the navigator hasn't finished mounting. See https://reactnavigation.org/docs/navigating-without-navigation-prop#handling-initialization for more details. ::


The solution is described pretty well here: https://reactnavigation.org/docs/navigating-without-navigation-prop/

Basically we need to wait for the ready state in addition to the mounted state.

We should send some of the error messages that happen when we really don't expect them to to the backend to analyze. Maybe we can sample 1% or something like that.

### Tested

end to end DeepLinkSend.spec.js

### Related issues

- Fixes #6434 

### Backwards compatibility

Yes.