Fix call to set gas currency in CLI base (#6505)
### Description

In the code to set cUSD as the gas currency for `celocli`, there is an error that sends the string
'StableToken' as an `Address` to `kit.updateGasPriceInConnectionLayer`. Instead, it should fetch the
address for the cUSD contract from the registry and pass that in, or we should update the
`kit.updateGasPriceInConnectionLayer` to take the contract identifier and complete the same
operation interanlly. This PR implements the former.

```
Error: invalid address (argument="address", value="StableToken", code=INVALID_ARGUMENT, version=address/5.0.4) (argument="tokenAddress", value="StableToken", code=INVALID_ARGUMENT, version=abi/5.0.0-beta.153)
Code: INVALID_ARGUMENT
```

### Tested

Ran a cUSD transfer command that previously failed