Bundle approve and request in a single action (#6431)
### Description

Implement the changes described in #6430 which removes the need to execute `cUSD.approve` as a separate transaction.
These changes are dependent on https://github.com/celo-org/komenci/pull/153 to be deployed in the API the client is pointing to. 

### Tested

Executed flow locally using the `attestation-flow` utility

### Related issues

- Fixes #6430 

### Backwards compatibility

Changes are _not_ backwards compatible.