[Wallet] Add native screen for Consumer Incentives (#6522)
### Description

Implement the Consumer Incentives screen natively. It was being shown through a WebView before.
The CTA button used to be a button that always linked to the cash in screen. Now, if the user is not verified, it will link to the phone verification screen.

<img width="344" alt="Screen Shot 2021-01-19 at 03 28 59" src="https://user-images.githubusercontent.com/6062888/104996344-80138a80-5a06-11eb-8dca-71248c954c99.png">

Note: Some copies are still pending.

### Other changes

Added a 'pill' to the drawer menu that links to the screen:
<img width="219" alt="Screen Shot 2021-01-19 at 03 46 54" src="https://user-images.githubusercontent.com/6062888/104997799-fadda500-5a08-11eb-8cbd-d8a69001103b.png">


### Tested

Manually and with unit tests.

### Related issues

- Fixes #6521

### Backwards compatibility

N/A