Add prepublish hook to generate the npm-shrinkwrap (#6575)
### Description

Fixed all the libraries that the cli is using at the moment of publishing it. 

### Tested

Published different versions of the cli under a different name, download the package and verified it.
Manually change the shrinkwrap to downgrade a version (package B), of a package required by a sub-package (package A) that is configurated to download the lastest version (of package B), and publish it again. The new published version downloaded the package B downgraded, instead of the lastest version


### Backwards compatibility

Yes