[Wallet]: Update radio button to new design  (#6501)
### Description

Current Behavior: The CheckCircle component is used to signify a selected item in the Settings page. This PR replaces the CheckCircle component with a file called RadioButton.tsx to better reflect the updated designs to be closer to what we have in the currentt design system.  

-Changes the CheckCircle.tsx file to RadioButton.tsx 
-Updates accordingly in files where CheckCircle was used 


### Other changes
Updates that RadioButton receives the "disabled" property from parent. This was an error that Cal caught! Fixed it up to make sure that the circle appears to be Disabled, rather than Unselected. 
Deletes circle styles defined in SelectionOption component 
Updates how RadioButton styles works 

### Tested


![image](https://user-images.githubusercontent.com/12700801/105538700-a8221880-5ca8-11eb-8b11-2b2067f9f932.png)

Will also test on Android 
### Related issues

- Fixes #6270

### Backwards compatibility

These are the only places so far that use this component, so it should be backwards compatible!