[Wallet] Add Profile Pictures (UI only) (#5970)
### Description

First, I know the PR looks big, and it is, but the number of lines changed is misleading, it's mostly snapshot files and a big file rename (`NameAndNumber` -> `NameAndPicture`).

Adds pictures to the Wallet. On the onboarding, user is not prompted to use their phone number until the verification step and they are asked for their name and a picture instead. They can change these two things on the "Edit Profile" option on settings. The image is shown in the drawer, but it's not used elsewhere yet and things are not being stored using CIP-8 yet.

### Other changes

I tried not to add unneeded things because of the size of the PR. I guess I could have left out the Profile screen changes for another PR.

### Tested

On physical iOS and Android devices.

Screenshots:
<img width="353" alt="Screen Shot 2021-01-28 at 14 36 37" src="https://user-images.githubusercontent.com/6062888/106176998-ba9fc480-6176-11eb-8be5-7b80f8fa823a.png">
<img width="345" alt="Screen Shot 2021-01-28 at 14 36 48" src="https://user-images.githubusercontent.com/6062888/106177009-bd9ab500-6176-11eb-92bc-61d6e822e069.png">
<img width="356" alt="Screen Shot 2021-01-28 at 14 37 02" src="https://user-images.githubusercontent.com/6062888/106177011-becbe200-6176-11eb-8175-df32ed93f372.png">
<img width="333" alt="Screen Shot 2021-01-28 at 14 37 44" src="https://user-images.githubusercontent.com/6062888/106177014-bffd0f00-6176-11eb-91e8-fc58035e31bd.png">
<img width="324" alt="Screen Shot 2021-01-28 at 14 38 08" src="https://user-images.githubusercontent.com/6062888/106177016-bffd0f00-6176-11eb-9086-5c4b41dfe795.png">
<img width="325" alt="Screen Shot 2021-01-28 at 14 38 14" src="https://user-images.githubusercontent.com/6062888/106177017-c095a580-6176-11eb-8052-bb3d4b308c55.png">
<img width="334" alt="Screen Shot 2021-01-28 at 14 39 33" src="https://user-images.githubusercontent.com/6062888/106177024-c12e3c00-6176-11eb-9ce6-007909874c25.png">



### Related issues

- Fixes #5969 & #4159

### Backwards compatibility

N/A