Fix Azure HSM Wallet PublicKey to Address conversion (#6829)
### Description

Some more context in issue #6807

Summary is that latest version of `sdk` (post-refactor) introduces some issues in how we interpret the Azure HSM KeyVault keys. Specifically how we convert the public key (EC x and y terms) into an ethereum address. 
This came to light when trying to run Komenci on staging based on the latest master and the issue exhibited itself as the service complaining that the wallet doesn't contain the address that it should.
This lead me to find that if I initialise a Wallet and I run `getAccounts()` using the same key vault but different versions of the codebase, it results in different accounts. But diving deeper into the code I see that the same public key is being manipulated.

I identified two separate issues in the flow which are causing the malformation.

#### 1. Using Buffer.from on arbitrary hex

In `azure-hsm-wallet.ts` we're converting a public key into an address by calling:
```typescript
const publicKey = await this.keyVaultClient!.getPublicKey(keyName)
return publicKeyToAddress(publicKey.toString(16))
```

Here `publicKey` is a BigNumber that's then turned into a hex string using `toString`. But this method has a little gotcha.
The public key is prefixed with `04` to inform that it is an uncompressed key. But because it's converted to a BigNumber the `0` at the front is actually dropped and we get something like `4<... other 128 chars / 64 bytes>` having a total of 129 chars. 

In `utils/src/address` the `publicKeyToAddress` function relied on `hexToBuffer` which is implemented as:
```typescript
export const hexToBuffer = (input: string) => Buffer.from(trimLeading0x(input), 'hex')
```

Here the usage of `Buffer.from` is a problem because with uneven hex strings it results in dropping a char at the tail:
```typescript
> Buffer.from('42236', 'hex')
<Buffer 42 23>
``` 

And because the above logic which uses BigNumbers to compute the public key, and the fact that there's `04` prefix, our keys are _always_ uneven and instead of being parsed as `<Buffer 04 .. .. .. ..>` we get `<Buffer 4. .. .. ..>` and we lose the last octet. 

In the PR I've reverted to using the `toBuffer` method from `ethereumjs-utils` because that properly handles this case, but we might also want to fix our own implementation. 

#### 2. Passing `isCompressed` to the `santize` parameter

The [pubToAddress](https://github.com/ethereumjs/ethereumjs-util/blob/master/src/account.ts#L251-L260) method takes an optional `sanitize` param that takes care of different formats in which a public key can be passed.

This is where I'm not sure what the intention was. From my current understanding we have two options here:

1. Always pass `true` for `sanitize` (as I did in the PR and was before on master).
2. Pass `isCompressed` but when `isCompressed` is false we need to also do a `trimUncompressedPrefix` to get rid of the `04` prefix.

(2) seams superfluous as that's exactly the point of `sanitize` here. But here I might be missing some info.

### Tested

This issue wasn't picked up by the specs because we were mocking the keyvault client in way that wasn't true to how it worked. We weren't returning a public key with the `04` prefix. I updated the tests to reflect that.

### Related issues

- Fixes #6807 

### Backwards compatibility

Fixes a regression.