Fix verifying the previous release on a network (#6850)
### Description

This PR fixes the [step from the docs that verifies the previous release on a network](https://docs.celo.org/community/release-process/smart-contracts#verify-the-previous-release-on-the-network)

The problem was that `CeloContractName` in `lib/registry-utils.ts` was being relied upon as an enumeration of all Celo Core Contracts. When new contracts are added later on, as of that commit, the tooling would fail as it does not know what to do with the new contract.

This intermediary fix for this problem is to copy over the target's `lib/registry-utils.ts` and effectively going forward treat that file as a source of truth. In the long-run, some more effective configuration of this should envisioned, as copying source code around is not ideal.


### Tested

- Ran `yarn verify-deployed -n mainnet -b celo-core-contracts-v1.mainnet -f` which previously failed

### Related issues

- Fixes #6615 

### Backwards compatibility

This PR effectively restores backwards compatibility, though may make it hard to do refactors in a backwards compatible way in the future.