Send custom notification title for CELO rewards (#6886)
### Description

When we send out rewards as part of the Consumer Incentives program, the notifications should now have a custom title (`Reward Received`).

### Other changes

N/A

### Tested

Manually. I commented out the code that sends the notifications, added logs to see notifications were attempted correctly and run the service on alfajores. We really need to improve the development process for this service.

### Related issues

- Fixes #6881

### Backwards compatibility

N/A