[Wallet] Addresses UI issue when user enters send amount larger than 100,000  (#6951)
### Description
Current behavior: When user enters an amount greater than 100,000 on the Send Amount screen, the amount breaks out of the line container and overlaps the screen. 
This PR addresses the issue by making the font a smaller size to fit the container as the amount gets larger. 

### Other changes

### Tested

Tested via iOS simulator. 
A sample of what it looks like when the font shrinks: 
![Simulator Screen Shot - iPhone 11 - 2021-02-05 at 14 40 33](https://user-images.githubusercontent.com/12700801/107097267-6bcedc00-67c1-11eb-9573-4d3d9b17494f.png)
What it looks like before the font gets smaller: 
![Simulator Screen Shot - iPhone 11 - 2021-02-05 at 14 40 28](https://user-images.githubusercontent.com/12700801/107097271-6ec9cc80-67c1-11eb-905c-653fd2545b10.png)


### Related issues

- Fixes #4960 
- https://github.com/celo-org/celo-monorepo/issues/4960

### Backwards compatibility

Yes