Increase block wait time to 600 on notification-service (#7029)
### Description

We have a cache of processed blocks so we query from `now-cache_size` to now. `cache_size`  used to be 120and this PR increases it to 600.

This is because we've seen missing notifications in periods of high load. Not sure if this will fix the issue, but it's unlikely to hurt either.

### Other changes

N/A

### Tested

N/A

### Related issues

N/A

### Backwards compatibility

N/A