Deduce providers names and add them to the feed (#3)
### Description

Transfers coming from Simplex and Moonpay are both routed through a Bittrex address causing them to appear to come from the same provider. This PR adds a mechanism to guess where the transaction comes from to include on the feed with a generic default option.

The solution I ended up going with is to save the last provider the user clicked on and if we then see an incoming tx from an address that we know belongs to a provider we assign it to the one the user clicked on. If a tx comes in from a provider address but the user didn't click on any provider, we show a generic 'cUSD Purchase' or 'CELO Purshase'.

To do this, I added an optional `isProviderAddress` field to the known addresses on Firebase.

### Other changes

- Added a new analytics event when we see a new tx that belongs to a provider.

### Tested

Locally, there are some unit tests as well.

### Related issues

- Fixes celo-org/celo-monorepo#7105

### Backwards compatibility

N/A