Use priority normal for android notifications (#48)
### Description

Still doing more testing with the flags but I believe we should use priority normal for android push notifications so our app does not get woken up in the background, which can cause other unintended consequences.

Looking at the ANR traces in sentry I think most of them happen while the app is in the background. The only way our app gets woken up in background is via high priority push notifications. Since devices are in low latency network mode whatever we attempt to send or connect to might fail. We do not actually need to do any processing in background at this point so I think it is better to not wake up the device at all. The system takes care of displaying the notification as usual and the app process gets created when the notification is tapped.

### Other changes

None

### Tested

curl firebase
https://celo-org.slack.com/archives/CL7BVQPHB/p1606917022075000

### Related issues

- Fixes celo-org/celo-monorepo#6542

### Backwards compatibility

Yes