Eliminate online gas estimation from send and invite flows (#117)
### Description

As described in https://github.com/celo-org/celo-monorepo/issues/5432, gas estimation is a costly
operation for the mobile client. The most promising short-term solution is to simply hard-code the
gas value to be used in for simple transactions along common code paths. The most common transaction
sent by a typical user is to send cUSD, either by a direct transfer or an invite.

This PR adds the required plumbing to use a single gas value, price and currency throughout the send
and invite flows. As a result, a single constant gas value can be used from start to finish. The
major result of this is that instead of estimating gas online twice, as the app currently does, the
static value is used and online gas estimation is never used.

This represents a large improvement to the experienced transaction speed, especially in less than
ideal network conditions. Additional effects of this are that errors caused by gas estimation should
be reduced, it's possible to give a fee estimate while offline, and the fee estimate shown the user
will be respected as a maximum fee value.

### Other changes

* Remove the unused `EstimateFee.tsx`
* Refactor cancellation for `sendTransaction`
* Fix gas price cache to handle multiple currencies.
* Propagate receipts back up from `sendTransaction`
* Use a Lock for getting contractkit instance to avoid reliance on polling and flag-based locking.
* Don't clear the transaction feed when the call to the blockchain API fails, such as when the device is offline.
* Add `networkConnectedSelector` to access the network connectivity state reported by the device.

### Tested

* Updated snapshot and unit tests to cover new functionality
* Manually tests effected user flows: Sending cUSD, escrow, and reclaiming escow

### Related issues

- Related: https://github.com/celo-org/celo-monorepo/issues/5432

### Backwards compatibility

No backwards compatibility concerns.